Move C++ command line utilities out of common
This utility file is specific to C++ and is
used only by the C++ extractor.

Partially addresses #112.
