+++ kythe/cxx/extractor/cxx_extractor.cc
@@ -39,11 +39,11 @@
 #include "absl/strings/string_view.h"
 #include "gflags/gflags.h"
 #include "glog/logging.h"
-#include "kythe/cxx/common/CommandLineUtils.h"
 #include "kythe/cxx/common/json_proto.h"
 #include "kythe/cxx/common/language.h"
 #include "kythe/cxx/common/path_utils.h"
 #include "kythe/cxx/common/proto_conversions.h"
+#include "kythe/cxx/extractor/CommandLineUtils.h"
 #include "kythe/proto/analysis.pb.h"
 #include "kythe/proto/buildinfo.pb.h"
 #include "kythe/proto/cxx.pb.h"
